pyproject.toml
[project]
name = "git-improved"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.21",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
